Update README.txt
A short C++ quiz program. Made it after watching the codeBeauty structures guide.
This is not my first code but it is the first one that took effort, even though it's only 80 lines.
I'm new to coding. This was done after 2 months from starting from 0.
I would love it if people just took a glance and left suggestions where I could improve or a general direction
I should head. At the moment I am keen on focusing on C++, but maybe in a year I could branch out to Java
for web stuff.