Add files via upload
Added the wall of text into a .txt file.
Made the program count all the lines in the .txt file and save integers in an array.
Got the program to print out the lines with a double loop incrementing both the start and end locations.
The challenge was to make it only print out 4 lines at a time (question and 3 answers) then adding the input functionality in between properly.

So yeah, kind of learned files. Took bashing my head a big but I got it.